refactor(store): decomposition ngxs module
